@@ -1,12 +1,19 @@
 require 'data_helpers'
+require 'elasticsearch'
 
 module Search
   class << self
     attr_reader :index_name
   end
 
-  def self.es
-    @es ||= Stretcher::Server.new('http://localhost:9200')
+  # TODO: Remove and replace with 'elasticsearch' stuffs.
+  # def self.es
+  #   @es ||= Stretcher::Server.new('http://localhost:9200')
+  # end
+
+  def self.client
+    @client ||= Elasticsearch::Client.new log: true
+    # @client.cluster.put_settings body: settings
   end
 
   # http://stackoverflow.com/questions/16205341/symbols-in-query-string-for-elasticsearch
@@ -62,7 +69,7 @@ module Search
   end
 
   def self.index
-    es.index(index_name)
+    client.index index: index_name
   end
 
   # Defines Elasticsearch index settings
@@ -132,73 +139,77 @@ module Search
   # A new index is created, populated and then aliased to the
   # main index name. The old index is deleted. This permits
   # zero downtime mapping changes
-  def self.migrate(index_suffix=nil)
-    # Find the previous index, if any
-    old_index = Search.true_index_name
+  # def self.migrate(index_suffix=nil)
+  #   # Find the previous index, if any
+  #   old_index = Search.true_index_name
 
-    # Create the new index, named by time of creation (or an argument)
-    # to avoid namespace conflicts with any old indexes
-    index_suffix = index_suffix || Time.now.to_i.to_s
-    new_index = "#{index_name}_#{index_suffix}"
-    puts "Creating new index #{new_index}"
-    es.index(new_index).create(settings: settings, mappings: mappings)
-    es.refresh
+  #   # Create the new index, named by time of creation (or an argument)
+  #   # to avoid namespace conflicts with any old indexes
+  #   index_suffix = index_suffix || Time.now.to_i.to_s
+  #   new_index = "#{index_name}_#{index_suffix}"
+  #   puts "Creating new index #{new_index}"
+  #   client.index index: new_index, body: mappings
+  #   client.indices.refresh
+  #   # es.index(new_index).create(settings: settings, mappings: mappings)
+  #   # es.refresh
 
-    # Check to make sure we actually need a new index here
-    unless old_index.nil?
-      old_settings = es.index(old_index).get_settings[old_index]['settings']
-      new_settings = es.index(new_index).get_settings[new_index]['settings']
+  #   # Check to make sure we actually need a new index here
+  #   unless old_index.nil?
+  #     # old_settings = es.index(old_index).get_settings[old_index]['settings']
+  #     # new_settings = es.index(new_index).get_settings[new_index]['settings']
+  #     old_settings = es.index(old_index).get_settings[old_index]['settings']
+  #     new_settings = es.index(new_index).get_settings[new_index]['settings']
 
-      # uuid and creation date always vary
-      old_settings['index']['uuid'] = new_settings['index']['uuid']
-      old_settings['index']['creation_date'] = new_settings['index']['creation_date']
+  #     # uuid and creation date always vary
+  #     old_settings['index']['uuid'] = new_settings['index']['uuid']
+  #     old_settings['index']['creation_date'] = new_settings['index']['creation_date']
 
-      old_mappings = es.index(old_index).get_mapping[old_index]['mappings']
-      new_mappings = es.index(new_index).get_mapping[new_index]['mappings']
+  #     old_mappings = es.index(old_index).get_mapping[old_index]['mappings']
+  #     new_mappings = es.index(new_index).get_mapping[new_index]['mappings']
 
-      if old_settings == new_settings && old_mappings == new_mappings
-        puts "Search settings/mappings are current, no migration needed"
-        es.index(new_index).delete
-        return
-      end
-    end
+  #     if old_settings == new_settings && old_mappings == new_mappings
+  #       puts "Search settings/mappings are current, no migration needed"
+  #       es.index(new_index).delete
+  #       return
+  #     end
+  #   end
 
-    # Populate the new index with data
-    Search.full_index(new_index)
+  #   # Populate the new index with data
+  #   Search.full_index(new_index)
 
-    if old_index.nil?
-      actions = []
-    else
-      actions = [
-        { remove: {
-          :alias => index_name,
-          :index => old_index
-        }}
-      ]
-      puts "Removing alias #{index_name} => #{old_index}"
-    end
+  #   if old_index.nil?
+  #     actions = []
+  #   else
+  #     actions = [
+  #       { remove: {
+  #         :alias => index_name,
+  #         :index => old_index
+  #       }}
+  #     ]
+  #     puts "Removing alias #{index_name} => #{old_index}"
+  #   end
 
-    puts "Adding alias #{index_name} => #{new_index}"
-    actions << [
-      { add: {
-        :alias => index_name,
-        :index => new_index
-      }}
-    ]
+  #   puts "Adding alias #{index_name} => #{new_index}"
+  #   actions << [
+  #     { add: {
+  #       :alias => index_name,
+  #       :index => new_index
+  #     }}
+  #   ]
 
-    es.request(:post, "_aliases", nil, actions: actions)
+  #   es.request(:post, "_aliases", nil, actions: actions)
 
-    unless old_index.nil?
-      puts "Deleting index #{old_index}"
-      es.index(old_index).delete
-    end
-  end
+  #   unless old_index.nil?
+  #     puts "Deleting index #{old_index}"
+  #     es.index(old_index).delete
+  #   end
+  # end
 
   # Force Elasticsearch to refresh the index
   # In practice this is done automatically, but the tests
   # need to know that the data is definitely available
   def self.refresh
-    @es.refresh
+    @client.indices.refresh
   end
 
   # Drop the current search index
@@ -206,13 +217,14 @@ module Search
   def self.drop(name=nil)
     name = name || index_name
     puts "Deleting index #{name}"
-    es.index(name).delete rescue nil
+    @client.indices.delete index: name
+    # es.index(name).delete rescue nil
   end
 end
 
 module Search::Paper
   def self.es_find(params)
-    res = Search.index.type(:paper).search(params)
+    res = Search.index.search body: params
     puts "  Elasticsearch (#{res.raw.took}ms) #{params}"
     res
   end
@@ -323,7 +335,7 @@ module Search::Paper
           author_ids = {}
           sciter_ids = {}
           paper = {
-            '_type' => 'paper',
+            # '_type' => 'paper',
             'uid' => row['uid'],
             'title' => row['title'],
             'abstract' => row['abstract'],
@@ -357,7 +369,12 @@ module Search::Paper
         papers[row['uid']]['feed_uids'] << row['feed_uid']
       end
 
-      result = Search.es.index(index_name).bulk_index(papers.values)
+      # result = Search.es.index(index_name).bulk_index(papers.values)
+
+      # TODO: This is totally wrong.
+      docs   = papers.values.map { |p| { "doc" => p } }
+      result = Search.client.bulk index: index_name, body: docs
+
       raise result if result.errors
 
       total += papers.length
